fix: reorganized code
